shift titleManager to ViewPrintDetailsForm.js + update test cases
