+++ src/settings/ViewPrintDetails/ViewPrintDetails.js
@@ -2,7 +2,7 @@ import { useIntl } from 'react-intl';
 import { useState } from 'react';
 
 import { Loading, Modal, Button, ModalFooter } from '@folio/stripes/components';
-import { TitleManager, useCallout } from '@folio/stripes/core';
+import { useCallout } from '@folio/stripes/core';
 
 import { usePrintDetailsSettings } from './hooks/usePrintDetailsSettings/usePrintDetailsSettings';
 import { usePrintDetailsSettingsMutation } from './hooks/usePrintDetailsSettingsMutation/usePrintDetailsSettingsMutation';
@@ -106,10 +106,7 @@ const ViewPrintDetails = () => {
   }
 
   return (
-    <TitleManager
-      page={formatMessage({ id: 'ui-circulation.settings.title.general' })}
-      record={formatMessage({ id: 'ui-circulation.settings.title.viewPrintDetails' })}
-    >
+    <>
       <ViewPrintDetailsForm
         onSubmit={onSubmit}
         key={reRenderFlag}
@@ -124,7 +121,7 @@ const ViewPrintDetails = () => {
       >
         {formatMessage({ id: 'ui-circulation.settings.ViewPrintDetails.warningPopupMessage' })}
       </Modal>
-    </TitleManager>
+    </>
   );
 };
